Refactor dom methods
