+++ src/domUpdates.js
@@ -1,3 +1,26 @@
+/* Query selectors */
+const dateFilterButton = document.querySelector('.date-filter-button');
+const loginButton = document.querySelector('#signInButton');
+const noRoomsMessage = document.querySelector('.no-rooms-message');
+const roomsContainer = document.querySelector('.room-options-container');
+const roomFilterButton = document.querySelector('.room-type-search-button');
+const roomFilterValue = document.querySelector('.room-type-search');
+const roomTypeCategories = document.querySelectorAll('.room-options');
+
+/* Render functions */
+const filterRoom = () => {
+  const roomTypeValue = document.querySelector('.room-type-input').value;
+  if (roomTypeValue === 'All') {
+    dashboard.filterRoomByType(roomTypeValue);
+    roomTypeCategories.forEach(cat => cat.classList.remove('hidden'));
+  } else {
+    roomTypeCategories.forEach(cat => cat.classList.add('hidden'));
+    document.querySelector(`.${roomTypeValue}-options`).classList.remove('hidden');
+  }
+  noRoomsMessage.classList.add('hidden');
+  dashboard.filterRoomByType(roomTypeValue);
+}
+
 const renderBookings = (guestBookings) => {
   guestBookings.map(booking => {
     renderBooking(booking);
@@ -20,8 +43,41 @@ const renderDashboard = (guestInfo) => {
   `);
 }
 
+const renderRooms = (roomTypes, rooms) => {
+  roomTypes.forEach(type => {
+    rooms.forEach(room => {
+      if (room.roomType === type) {
+        document.querySelector(`.${type.split(' ').join('-')}-container`).insertAdjacentHTML('afterEnd', `
+          <div class="room-div">
+            <p>Room Number: ${room.number}</p>
+            <p>Bed Size: ${room.bedSize}</p>
+            <p>Number of Beds: ${room.numBeds}</p>
+            <button value="${room.number}" class="book-button">Book Room<buttton/>
+          </div>
+        `)
+      }
+    })
+  })
+}
+
+const renderTotalSpent = (totalSpent) => {
+  document.querySelector('.total-spent').insertAdjacentHTML('beforeEnd', `
+  <p>You have spent a total of $${totalSpent} with us!</p>
+`)
+}
+
 export {
+  dateFilterButton,
+  filterRoom,
+  loginButton,
+  noRoomsMessage,
+  roomsContainer,
+  roomFilterButton,
+  roomFilterValue,
+  roomTypeCategories,
   renderBooking,
   renderBookings,
   renderDashboard,
+  renderRooms,
+  renderTotalSpent
 }
